requirements.txt
flask
flask-smorest
python-dotenv
requests
selenium
webdriver_manager
beautifulsoup4
marshmallow
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
line-bot-sdk
apscheduler
pandas
html5lib